base: Force specific package versions for upstream packages
Force the current latest versions of the python packages installed
in the SWAN base image, except for the notebook one. This package
is using the latest version (6.5.6) before major version 7, as there
are incompatibility issues that arise due to the usage of such
version along with the usage of nbclassic or swanclassic packages.

This is due to the fact that having the notebook and nbclassic/
swanclassic packages installed results in the base path of swanclassic
changing and it is assumed that it is exposed in the root path.