swan: Move CUSTOM_JS_DIR and KERNEL_DIR variables to Dockerfile
These variables do not need to be defined in run time and can be
defined in build time. This also applies for the instructions
that create the directories based on such variables.

Furthermore, created custom.js file and added it to the Dockerfile
instead of echoing the javascript code inside during run time.