swan: Remove unecessary instructions from the SWAN Dockerfile
The NB_USER arg is no longer required in the Dockerfile and thus it was
removed.

Furthermore, remove instructions that disabled requiretty and
secure_path, as there is no indication that this is needed.

Also removed the instruction that appened the texlive binary to the
user PATH, as texlive is now installed through DNF.