swan: Add script files to docker container
Add the shell script files to the docker container, that will run on
container startup to configure the storage, cvmfs/sofware environment
and the user environment.

Additional files and scripts are added, which were previously created
and populated in the bash scripts that were executed upon container
startup, in the previous centOS image.