swan: Change file authors and use new log function
Update the authors of the shell scripts to the current SWAN team.
Each shell script has a number appended to their name so that they
are executed in the same and correct order, due to the dependencies
between them.

Remove the log_info and log_error functions and use the upstream one.
For the 03_userconfig.sh script we had to define a new function, as it
is ran as the user of the container instead of root.