swan-cern: Do not exit bash script if env vars are not defined
This was changed because the bash scripts are sourced and so, when they
would exit, that would cause the main start.sh script to exit as well,
leading to halting the user container.

Furthermore, changed the user group of the user of the container, as it
is defined by NB_GID environment variable.