swan,swan-cern: Create a new nb_term_lib directory for user-exposed libraries
This commit introduces a convention by creating a new directory
where to install libraries that will be included in the user
environment for notebooks and terminals. swandaskcluster will
be installed in this directory from now on, instead of having
a dedicated DASK_LIB_DIR.

We also create a SWAN_LIB_DIR variable that will prefix the
aforementioned directory (SWAN_LIB_DIR/nb_term_lib), as well as
the already existing directory to expose Jupyter extensions
(SWAN_LIB_DIR/extensions).