swan: Prevent TERM=unknown from being injected in the terminal environment
When running configure_kernels_and_terminal.py as the user, the
TERM variable has "unknown" as value, which interferes with the
correct usage of the terminal. Here we prevent such value from
being injected, which makes it possible that the terminal shows
the TERM=xterm-256 when finally working in JupyterLab. This value
is somehow inherited from base images (up to the Alma9 one).