base: Disable warnings about frozen modules
Python3.11 can freeze its core modules for performance reasons.
IPython seems to import a module which triggers a warning about
not being able to debug those core modules because they are frozen.

Here we disable such warning for the Jupyter server, which runs
on Python3.11. No debugging of the core modules of the server is
expected and, if so, we could run it with frozen_modules=off if
necessary in a dev environment.

More information here:
- https://stackoverflow.com/questions/75114841/debugger-warning-from-ipython-frozen-modules
- https://discourse.jupyter.org/t/debugger-warning-it-seems-that-frozen-modules-are-being-used-python-3-11-0/16544/5