swan-cern: Fix htcondor script to disable dask if HTCondor is disabled
If CERN_HTCONDOR is set to false, the dask-labextension is supposed to
be disabled. Therefore, we add it to a json object along with the
sparkconnector and sparkmonitor extensions if spark is not enabled as
well. Otherwise, the json object is created and the dask-labextension is
added to the diasbled extensions json object