swan: Remove instruction that copies jupyter configuration from the LCG releases
This instruction was removed to prevent loading jupyter configuration
from the LCG releases, so that we are not dependent on modules and
extensions that the LCG release tries to activate or disable in the
jupyter config.

The nbextensions configurator extension needs to be enabled now because it
was done before by one of the copied config files, and it's a package that
is installed in the swan image.