base: Remove dnf upgrade instruction
The instruction was removed as we are using a fixed version of the alma9
image, so that the layers on top are based on a stable build. By running
the dnf upgrade instruction we were doing the same thing as if we were
always using the latest version of the alma9-base image, thus having no
stability.