swan-cern: Create symlink to voms-proxy-init on grid.cern.ch
So that users can create their proxies to external data sources
or the grid more easily. Using a voms-proxy-init from CVMFS
also prevents the installation of the tool in the image, which
we want to keep thin.

To be replaced when the corresponding Alma9 voms-proxy-init is
available on /cvmfs/grid.cern.ch.