New acc-py image and LCG prevention
The new images is built on top of `swan-cern` and install 3 different
version of `acc-py`, `java-11-openjdk` and a temporary installation
and configuration of `SwanCustomEnvironments` (to be removed in future).

When custom environments are launched, it is intended to avoid
extra packages and installations, including avoiding LCG
configurations. This means if the environment variables regarding
LCG are not set, the LCG configurations on `02_environment.sh`