swan-cern: Install cronie
With the addition of cronie, the CVMFS prefetcher container of
the CVMFS-CSI chart can use the same image as the user container.
This is useful because we make sure that the source and python
commands executed by the prefetcher run without errors in the same
environment as the user (image, LCG views from CVMFS).