[swan] Update jupyter-resource-usage extension to v1.1.0
The current version, 1.0.1, triggers the following responses:
403 GET /user/username/api/metrics/v1

when trying to get the resource metrics in the classic UI, since
the extension is not adapted to the required xsrf headers in
jupyterhub v4.1.6.

Updating to v1.1.0 does not make the extension compatible with
hub v4.1.6 in the classic UI, but at least it does not show
anymore the 403 messages. It does not show either the Memory
consumption in the classic UI, but the Lab UI is not affected.