Revert "swan: Set PIP_TARGET env variable"
This reverts commit 6cd7b28f087ec8e256017162da74b5e1bd99c5b2.

When a user uses --user flag, pip errors out as we already force the
usage of the --target flag with the PIP_TARGET variable. Moreover, when
a session is started with custom environments, the packages should not
be installed by default in the user CERNBox (i.e. in the site-packages
in their CERNBox).