swan: [SWAN-410] Allow editable packages to be found by Python
PYTHONNOUSERSITE is set to 1 in the base image to prevent the
Jupyter server process, notebook processes and terminal processes
from inspecting the user site packages directory by default.

However, if PYTHONNOUSERSITE is set to 1, editable packages are
ignored by Python (see SWAN-410 for more details). Therefore,
if the user selects to work with packages installed on CERNBox
(editable or not), we set PYTHONNOUSERSITE to 0 to prevent
breaking the editable package search mechanism.