swan: Add env variables to include extra PATH directories
When running a script as sudo and as another user, the PATH and
PYTHONPATH variables are not propagated to the user environment.
Moreover, this is intended, as the user environment should use the
python and libraries coming from the LCG release or that are installed
by the users themselves.

However, there may be extensions installed in the swan images, which
require specific paths to be included in the aforementioned variables,
in the user environment. For this, USER_PATH and USER_PYTHONPATH have
been introduced, which are then appended to the other two variables and
seen by the user in the terminal and notebooks