feat: integration of Rucio Jupyter extension in SWAN
This commit introduces two main changes to the image: a configuration script
and an updated Dockerfile.

The script performs the followign operations:

1. Sets ENV vars such as RUCIO_HOME, PATH, and PYTHONPATH based on the specified RUCIO_JUPYTERLAB_VERSION,
and validates that the selected version is available.

2. Configures the default iPython kernel to load the rucio_jupyterlab extension.

3. Appends the Rucio-specific settings (from CVMFS) to the Jupyter server configuration file.

On the other hand, the Dockerfile includes:

1. Setting of the RUCIO_JUPYTERLAB_VERSION, so that it's accessible in the runtime scripts.

2. Installation of the Rucio extension.

3. Symlink of the extension's dependencies (living in CVMFS)

For the sake of completeness, documentation on how to use the extension is being prepared in https://github.com/swan-cern/help/pull/37.

Co-Authored-By: Enrique Garcia <enrique.garcia.garcia@cern.ch>